requirements.txt
torch
torchvision
opencv-python
scikit-learn
numpy
matplotlib
Pillow